+++ apps/web/src/app/customer/_shared/api/get-proposal-list.ts
@@ -1,7 +1,7 @@
 import { API_BASE_URL } from "@/shared/api/config";
 import { fetchJson } from "@/shared/api/fetch-json";
-import { proposalListSchema } from "../_model/proposal.schema";
-import type { ProposalList } from "../_model/proposal.schema";
+import { proposalListSchema } from "../model/proposal.schema";
+import type { ProposalList } from "../model/proposal.schema";
 
 export function getProposalList(
   reportId: string,
