+++ apps/web/src/app/customer/_shared/model/proposal.schema.ts
